+++ ModelSwapperSkins/BoneMapping/ModelSkinExclusionRuleType.cs
@@ -1,0 +1,8 @@
+﻿namespace ModelSwapperSkins.BoneMapping
+{
+    public enum ModelSkinExclusionRuleType
+    {
+        ExcludeIfApplied,
+        ExcludeIfNotApplied
+    }
+}